adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = ""
authors = [
    {name = "SimHoZebs",email = "simhozebs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-adk (>=1.15.1,<2.0.0)",
    "requests (>=2.0.0,<3.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
